requirements.txt
bertopic
sentence-transformers
scikit-learn
hdbscan
umap-learn
numpy
openai
python-dotenv
fastapi
uvicorn[standard]
pydantic-settings
nltk
